requirements.txt
httpx
aiosqlite
PyJWT